Bunch of random changes to GPSTest4
Didn't want to overwrite the current GPSTest4 as I did a bunch of stuff
in this guy, it works, just lots of changes.

Added a bunch of comments, lots and lots of commenting. Anything with
layers of stars around it is very important for us. Left some comments
on things we could use for other parts of the project (timer mainly)

Started commenting out more lines and funcs we won't be using. Think
I've scrapped most of them this time around.

Started modifying the funcs we will be using.

Update how the maps are displayed, getting it primed to be put on our
dashboard, main change was altering how/when the realtime path line is
drawn.

Changed a bit of the math.

haven't started on the CoreData stuff yet.

I think I'm missing some stuff, if I am it will be in the comments.

If you guys have any questions about the comments I left, feel free to
ask me. I'm too tired atm to go back and remove some of my working
comments... or some of the more vulgar comments when I was getting
frustrated at stuff. Will fix sometime tomorrow.